+++ squigglepy/distributions.py
@@ -42,6 +42,9 @@ def tdist(x, y, t, lclip=None, rclip=None):
 def log_tdist(x, y, t, lclip=None, rclip=None):
     return [x, y, 'log-tdist', t, lclip, rclip]
 
+def triangular(left, mode, right, lclip=None, rclip=None):
+    return [left, mode, 'triangular', right, lclip, rclip]
+
 def exponential(scale):
     return [scale, None, 'exponential', None, None]